Cleaned up some stuff that was for test or just unsed.
