+++ GardenModules/README.md
@@ -1,0 +1,3 @@
+# Garden Modules
+
+This folder houses the gardent modules. Each module represents and manages some form of input or output, e.g. a soil moisture sensor. The idea behind this approach is to make adding features as simple as adding a new module.
