--- ModuleTests/Data/luxData.csv
@@ -1,3 +1,0 @@
-527.978462833967,2020-11-29 02:55:51.918946
-528.5072462839585,2020-11-29 02:56:02.070627
-529.5646851648721,2020-11-29 02:56:12.221543
