+++ Tests/README.md
@@ -1,0 +1,3 @@
+# Module Tests
+
+This folder is used to test new modules or just anything to do with the garden.
