+++ Tests/camera_test.py
