--- rc.local
@@ -1,25 +1,0 @@
-#!/bin/sh -e
-#
-# rc.local
-#
-# This script is executed at the end of each multiuser runlevel.
-# Make sure that the script will "exit 0" on success or any other
-# value on error.
-#
-# In order to enable or disable this script just change the execution
-# bits.
-#
-# By default this script does nothing.
-
-# Print the IP address
-_IP=$(hostname -I) || true
-if [ "$_IP" ]; then
-  printf "My IP address is %s\n" "$_IP"
-fi
-exec 2> /home/pi/rc_local_logs.log
-exec 1>&2
-set -x
-sudo ./home/pi/go/src/github.com/papertrail/remote_syslog2/build/remote_syslog2/remote_syslog -p 35718 -d logs4.papertrailapp.com --pid-file=/var/run/remote_syslog.pid /home/pi/Desktop/smartGarden/smartGarden/logs/smartGardenLog.log &
-sudo python3 /home/pi/Desktop/smartGarden/smartGarden/smartGarden.py &>> /home/pi/Desktop/smartGarden/smartGarden/smartGardenLog.txt &
-#sudo /home/pi/go/src/github.com/elastic/beats/filebeat/filebeat -c /home/pi/go/src/github.com/elastic/beats/filebeat/filebeat.yml -e &
-exit 0
